Fix type of length returned by gvdb_table_get_names()
It should not be unsigned. The type in the on-disk format is gint32, so
we need to return something at least as wide as that. However, we
should not expose the implementation detail that the on-disk format is
specifically gint32. Use a gsize, since that’s the normal type for array
lengths — but check that we’re not on a platform where (somehow) gsize
is smaller than gint32.

+++ gvdb-reader.h
@@ -38,7 +38,7 @@ G_GNUC_INTERNAL
 void                    gvdb_table_free                                 (GvdbTable    *table);
 G_GNUC_INTERNAL
 gchar **                gvdb_table_get_names                            (GvdbTable    *table,
-                                                                         gint         *length);
+                                                                         gsize        *length);
 G_GNUC_INTERNAL
 gchar **                gvdb_table_list                                 (GvdbTable    *table,
                                                                          const gchar  *key);
